Recovered from Eclipse.
git-svn-id: https://subversion.assembla.com/svn/cmosher@6615 884d43a4-ffbe-47e9-90ca-fd8c37d7eb32

+++ DosMasterToImage/src/DosMasterToImage.java
@@ -44,10 +44,11 @@ public class DosMasterToImage
 
         in.mark(in.available());
 
-        int c = in.skip(0x1B00);
+        long c = in.skip(0x1B00);
         if (c != 0x1B00)
         {
             throw new RuntimeException("error skipping bytes");
+
         }
         for (int i = 0x3600; i < 0x4000; ++i)
         {
